python/pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "clevertools"
version = "1.4.2"
description = "Compact Python utility library for file I/O, config loading, sensitive-value masking, and ready-to-use logging."
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"
license-files = ["LICENSE", "NOTICE", "CONTRIBUTORS"]
authors = [{ name = "b7binw13" }]
keywords = ["utilities", "logging", "config", "file_handling", "json", "toml", "yaml", "masking", "b7binw13", "tbl", "the-binary-labs(tbl)"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "PyYAML>=6.0.3",
    "tomli_w>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "Faker>=40.11.0",
    "iniconfig>=2.3.0",
    "librt>=0.8.1",
    "mypy>=1.19.1",
    "mypy_extensions>=1.1.0",
    "packaging>=26.0",
    "pathspec>=1.0.4",
    "pluggy>=1.6.0",
    "Pygments>=2.19.2",
    "pytest>=9.0.2",
    "ruff>=0.15.7",
    "typing_extensions>=4.15.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/The-Binary-Labs-TBL/clevertools"
Repository = "https://github.com/The-Binary-Labs-TBL/clevertools.git"
Issues = "https://github.com/The-Binary-Labs-TBL/clevertools/issues"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]